requirements.txt
google-generativeai
python-dotenv
fpdf2
